Use the new recording model for the add recording screen

+++ app/src/main/java/org/tvheadend/tvhclient/fragments/recordings/RecordingDetailsFragment.java
@@ -331,7 +331,7 @@ public class RecordingDetailsFragment extends DialogFragment implements HTSListe
                 if (rec != null) {
                     DialogFragment editFragment = RecordingAddFragment.newInstance();
                     Bundle bundle = new Bundle();
-                    bundle.putLong("dvrId", rec.id);
+                    bundle.putInt("dvrId", rec.id);
                     editFragment.setArguments(bundle);
                     editFragment.show(activity.getSupportFragmentManager(), "dialog");
                 }
