+++ app/src/main/java/org/tvheadend/tvhclient/fragments/recordings/RecordingListFragment.java
@@ -140,7 +140,7 @@ public class RecordingListFragment extends Fragment implements HTSListener, Frag
             // Create the fragment and show it as a dialog.
             DialogFragment editFragment = RecordingAddFragment.newInstance();
             Bundle bundle = new Bundle();
-            bundle.putLong("dvrId", adapter.getSelectedItem().id);
+            bundle.putInt("dvrId", (int) adapter.getSelectedItem().id);
             editFragment.setArguments(bundle);
             editFragment.show(activity.getSupportFragmentManager(), "dialog");
             return true;
@@ -252,7 +252,7 @@ public class RecordingListFragment extends Fragment implements HTSListener, Frag
             // Create the fragment to edit a recording but show it as a dialog.
             DialogFragment df = RecordingAddFragment.newInstance();
             Bundle bundle = new Bundle();
-            bundle.putLong("dvrId", rec.id);
+            bundle.putInt("dvrId", (int) rec.id);
             df.setArguments(bundle);
             df.show(activity.getSupportFragmentManager(), "dialog");
             return true;
